requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-computedfields==0.2.1
django-fast-update==0.2.1
djangorestframework==3.14.0
drf-spectacular==0.26.0
inflection==0.5.1
jsonschema==4.17.3
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
